Updates from code review meetings DEC2018-FEB2019
We've been combing through an example workflow/data for basic 1H 1D NMR analysis. Each function that is called in that workflow is being updated for consistency/functionality.
There is one major change: ref_spectra now includes an extra (optional) argument that allows the user to see clearly whether the threshold is what it ought to be. This requires all standard parameters to be provided, but should not interfere with previous calls to this function.
MJ 9FEB2019